Lab 6/requirements.txt
# ══════════════════════════════════════════════
# Lab 6 — Python Dependencies
# Install: pip install -r requirements.txt
# ══════════════════════════════════════════════

# Web framework
fastapi==0.111.0
uvicorn[standard]==0.29.0

# HTTP client (async)
httpx==0.27.0

# Async file I/O
aiofiles==23.2.1

# Data validation
pydantic==2.7.1

# Environment variables
python-dotenv==1.0.1

# Rate limiting (Lab 6 — gateway)
slowapi==0.1.9

# Testing
pytest==8.2.0
pytest-asyncio==0.23.6
